Add core and rearrange types
