+++ src/computation.ts
@@ -1,3 +1,4 @@
+import type { ClaimedTask, Task } from "resonate-inner";
 import type { Clock } from "./clock";
 import { InnerContext } from "./context";
 import { Coroutine, type LocalTodo, type RemoteTodo } from "./coroutine";
@@ -7,7 +8,7 @@ import type { CallbackRecord, DurablePromiseRecord, Network } from "./network/ne
 import { Nursery } from "./nursery";
 import { AsyncProcessor, type Processor } from "./processor/processor";
 import type { Registry } from "./registry";
-import type { Callback, ClaimedTask, Func, Task } from "./types";
+import type { Callback, Func } from "./types";
 import * as util from "./util";
 
 export type Status = Completed | Suspended;
