+++ src/coroutine.ts
@@ -1,8 +1,8 @@
-import type { Context, InnerContext, Yieldable } from "./context";
-import { Decorator } from "./decorator";
+import type { Context, InnerContext } from "./context";
+import { Decorator, type Literal, type Value } from "./decorator";
 import type { Handler } from "./handler";
 import type { DurablePromiseRecord } from "./network/network";
-import { type Callback, type Literal, type Result, type Value, ko, ok } from "./types";
+import { type Callback, type Result, type Yieldable, ko, ok } from "./types";
 import * as util from "./util";
 
 export type Suspended = {
