+++ src/index.ts
@@ -1,7 +1,4 @@
-export { Resonate } from "./resonate";
+export { Resonate, ResonateHandle, ResonateFunc } from "./resonate";
 export { Context } from "./context";
-export * from "./types";
-export * from "./promises";
-export * from "./schedules";
-export * from "./tasks";
-export * from "./decorator";
+
+export * as core from "./core";
